requirements.txt
beautifulsoup4==4.11.1
bs4==0.0.1
cachetools==5.2.0
certifi==2022.12.7
chardet==3.0.4
charset-normalizer==2.1.1
dnspython==2.2.1
et-xmlfile==1.1.0
google-api-core==2.11.0
google-auth==2.15.0
google-auth-oauthlib==0.8.0
google-cloud-language==2.7.0
googleapis-common-protos==1.57.1
googletrans==4.0.0rc1
grpcio==1.51.1
grpcio-status==1.51.1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
numpy==1.24.1
oauthlib==3.2.2
openai==0.25.0
openpyxl==3.0.10
pandas==1.5.2
pandas-stubs==1.5.2.230105
proto-plus==1.22.1
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
pymongo==4.3.3
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.7
requests==2.28.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
tqdm==4.64.1
types-pytz==2022.7.0.0
typing_extensions==4.4.0
urllib3==1.26.13
